20.01 Update: rename DHTlib into DHT11
